Removes unused services
