+++ src/ipc-events.ts
@@ -63,13 +63,10 @@ export enum IpcEvents {
 
   FILE_GET_CONTENT = 'FILE_GET_CONTENT',
   IGNORED_FILES = 'IGNORED_FILES',
-  UNIGNORED_FILES = 'UNIGNORE_FILES',
   FILE_GET = 'FILE_GET',
   FILE_GET_ID_FROM_PATH = 'FILE_GET_ID_FROM_PATH',
 
   RESULTS_GET = 'RESULTS_GET',
-  RESULTS_GET_NO_MATCH = 'RESULTS_GET_NO_MATCH',
-
 
   EXPORT = 'EXPORT',
   EXPORT_NOTARIZE_SBOM = ' EXPORT_NOTARIZE_SBOM',
@@ -140,12 +137,9 @@ export const ipcRendererEvents = [
   IpcEvents.FILE_GET_CONTENT,
   IpcEvents.FILE_GET,
   IpcEvents.IGNORED_FILES,
-  IpcEvents.UNIGNORED_FILES,
   IpcEvents.FILE_GET_ID_FROM_PATH,
 
   IpcEvents.RESULTS_GET,
-  IpcEvents.RESULTS_GET_NO_MATCH,
-
 
   IpcEvents.EXPORT,
   IpcEvents.EXPORT_NOTARIZE_SBOM,
