+++ src/main/services/LogicInventoryService.ts
@@ -22,7 +22,7 @@ class LogicInventoryService {
   public async detach(inv: Partial<Inventory>): Promise<boolean> {
     try {
       const project = workspace.getOpenedProjects()[0];
-      await project.scans_db.results.restore(inv.files);
+      await project.scans_db.files.restore(inv.files);
       await project.scans_db.inventories.detachFileInventory(inv);
       const emptyInv: any = await project.scans_db.inventories.emptyInventory();
       if (emptyInv) {
